moved get_dummies into reshape module from utils.utils
