+++ python/cudf/reshape/general.py
@@ -147,3 +147,69 @@ def melt(frame, id_vars=None, value_vars=None, var_name=None,
         index=None)
 
     return DataFrame(mdata)
+
+
+def get_dummies(df, prefix='', prefix_sep='_', dummy_na=False, columns=None,
+                cats={}, sparse=False, drop_first=False, dtype='float64'):
+    """ Returns a dataframe whose columns are the one hot encodings of all
+    columns in `df`
+
+    Parameters
+    ----------
+    df : cudf.DataFrame
+        dataframe to encode
+    prefix : str, dict, or sequence, optional
+        prefix to append. Either a str (to apply a constant prefix), dict
+        mapping column names to prefixes, or sequence of prefixes to apply with
+        the same length as the number of columns. If not supplied, defaults
+        to the empty string
+    prefix_sep : str, optional
+        separator to use when appending prefixes
+    dummy_na : boolean, optional
+        Right now this is NON-FUNCTIONAL argument in rapids.
+    cats : dict, optional
+        dictionary mapping column names to sequences of integers representing
+        that column's category. See `cudf.DataFrame.one_hot_encoding` for more
+        information. if not supplied, it will be computed
+    sparse : boolean, optional
+        Right now this is NON-FUNCTIONAL argument in rapids.
+    drop_first : boolean, optional
+        Right now this is NON-FUNCTIONAL argument in rapids.
+    columns : sequence of str, optional
+        Names of columns to encode. If not provided, will attempt to encode all
+        columns. Note this is different from pandas default behavior, which
+        encodes all columns with dtype object or categorical
+    dtype : str, optional
+        output dtype, default 'float64'
+    """
+    if dummy_na:
+        raise NotImplementedError("dummy_na is not supported yet")
+
+    if sparse:
+        raise NotImplementedError("sparse is not supported yet")
+
+    if drop_first:
+        raise NotImplementedError("drop_first is not supported yet")
+
+    from cudf.multi import concat
+    if columns is None:
+        columns = df.columns
+
+    if isinstance(prefix, str):
+        prefix_map = {}
+    elif isinstance(prefix, dict):
+        prefix_map = prefix
+    else:
+        prefix_map = dict(zip(columns, prefix))
+
+    return concat([
+        df.one_hot_encoding(
+            name,
+            prefix=name
+            + (prefix_sep if prefix else '')
+            + prefix_map.get(name, prefix),
+            cats=cats.get(name, df[name].unique()),
+            prefix_sep=prefix_sep,
+            dtype=dtype)
+        for name in columns
+    ], axis=1)
